Created FhirParserHelper, moved demographic mapping method, created FhirParserTests project
